Add GCS bucket for terraform state
Terraform could be handle in a bucket from Google Cloud Storage.
Nevertheless, it leads to a 'chicken or the egg' problem:
a dedicated infrastructure should be used to deploy bucket,
then this bucket should use itself as its backend,
and then an other infrastructure can be bootstrap with that backend.
Terraform can be handle with Terraform cloud to avoid those issues.